+++ database/migrations/2025_05_05_202945_add_balance_to_users_table.php
@@ -8,7 +8,7 @@ return new class () extends Migration {
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->decimal('balance', 15, 2)->default(0.00);
+            $table->integer('balance')->default(0);
         });
     }
